refactor: extract sendResponse to avoid instantiating fake command classes
