+++ src/lib/interaction.ts
@@ -13,6 +13,27 @@ import RaidCommand from "../interactions/commands/raid";
 import Makibot from "../Makibot";
 import InteractionCommand from "./interaction/basecommand";
 import logger from "./logger";
+import axios from "axios";
+
+export async function sendResponse(event: APIGuildInteraction, response: string, ephemeral: boolean = false): Promise<void> {
+  const payload: any = {
+    type: 4,
+    data: { content: response },
+  };
+  if (ephemeral) {
+    payload.data.flags = 64;
+  }
+  logger.debug("[interactions] sending response: ", payload);
+  return axios.post(
+    `https://discord.com/api/v8/interactions/${event.id}/${event.token}/callback`,
+    payload,
+    {
+      headers: {
+      "Content-Type": "application/json",
+    },
+  },
+  );
+}
 
 interface HandlerConstructor {
   // https://stackoverflow.com/a/39614325/2033517
